changing sessionId from int to long

+++ drools-core/src/main/java/org/drools/command/SingleSessionCommandService.java
@@ -19,6 +19,6 @@ package org.drools.command;
 import org.drools.runtime.CommandExecutor;
 
 public interface SingleSessionCommandService extends CommandService {
-	int getSessionId();
+	long getSessionId();
 	void dispose();
 }
